use task logger extensively

+++ rpc/core/RpcResult.java
@@ -99,7 +99,7 @@ class RpcResult {
 	}
 
 	public
-	int getHttpStatus () {
+	long getHttpStatus () {
 
 		return Rpc.statusToHttpStatus (
 			status);
